implement `provider-interop^0.2`
I chose to make the container act as service-provider - I could have
chosen to make the container-factory do that, or even both of them.

However, because this is an immutable container-implementation, and
since creating the actual container is practically zero overhead in
this implementation, it made more sense this way: it ensures that the
internal entry-table is fully populated and immutable by the time the
container gets added as a provider to an external service-registry.